+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 0.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix object retrieving in portlets and related items.
+  Now expired items are not showed. (#108)
+  [cekk]
 
 
 0.1.0 (2016-01-05)